Updated demo app, APK, as well as README.
