+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.huhx0015.huhxagse"
+    package="com.huhx0015.hxaudiodemo"
     android:installLocation="auto">
 
     <application
@@ -10,13 +10,15 @@
         android:theme="@style/AppTheme" >
 
         <activity
-            android:name=".activity.HXGSEDemoActivity"
-            android:label="HX Game Sound Engine Demo"
+            android:name="com.huhx0015.hxaudiodemo.activity.HXAudioPlayerDemoActivity"
+            android:label="@string/app_name"
             android:configChanges="orientation|screenSize|keyboardHidden">
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+
         </activity>
 
     </application>
