+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':hxgse'
+include ':app', ':hxaudio'
